Master Dissertation
Master Dissertation solution for MSc Data Science. This solution is a python based deep learning algorithms that attempt to sentiment analysis for twitter data.
the BERT model has been developed for two different datasets. first data is a small data and related to the APPLE company and second data is bigger and 
related to seven big tech company on twitter.